requirements.txt
Django==6.0.2
pytest==9.0.2
pytest-django==4.11.1
pytest-cov==6.0.0
pytest-html==4.1.1
factory-boy==3.3.3
Faker==36.1.1
